# [Performance] Faster `SliceSampler._tensor_slices_from_startend`