## Description

Speeds up the `SliceSampler._tensor_slices_from_startend` method by about 8x in the case where `seq_length` is an int.

Running the performance measurement script from https://github.com/pytorch/rl/issues/2422#issue-2508402938 in my machine gives:

```
Without SliceSampler: 0.00019115543303390345 s
With SliceSampler: 0.002860310899753434 s
Slowdown factor: 14.963272842190824x
```

whereas the output before the change was:

```
Without SliceSampler: 0.0001870056662786131 s
With SliceSampler: 0.0046725646670286855 s
Slowdown factor: 24.98621972270721x
```

So this change provides a speedup of about `(0.00467 / 0.00286) = 1.632` to the `ReplayBuffer.sample` method for the particular case in that script.

I also took a performance profile of the script with `cProfile`, like so:

```bash
pthon -m cProfile <script> | grep _tensor_slices_from_startend
```

I increased the `timeit` iterations from 30 to 3000 for better precision. Before the change, the cumulative time spent in the `_tensor_slices_from_startend` function was 7.571. After the change, it was 0.871, so the speedup for `_tensor_slices_from_startend` alone was about 8x.

## Motivation and Context

close #2422

- [x] I have raised an issue to propose this change ([required](https://github.com/pytorch/rl/issues) for new features and bug fixes)

## Types of changes

What types of changes does your code introduce? Remove all that do not apply:

- [x] Bug fix (non-breaking change which fixes an issue)

## Checklist

Go over all the following points, and put an `x` in all the boxes that apply.
If you are unsure about any of these, don't hesitate to ask. We are here to help!

- [x] I have read the [CONTRIBUTION](https://github.com/pytorch/rl/blob/main/CONTRIBUTING.md) guide (**required**)