@kurtamohler 2024-09-09
> IIRC this https://github.com/kurtamohler/torchrl/blob/db5f5cff8a67f3854759ab78215046bc65019046/torchrl/data/replay_buffers/samplers.py#L1881 used to be the most expensive thing when the buffer is full (eg 1M elements). Can you reproduce that in your benchmark? If so a follow-up should be to speed up that one too!

I see, I will look into that